auto refresh movies after search
