+++ src/components/SearchResult.js
@@ -1,10 +1,38 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './SearchResult.css';
+import axios from 'axios';
 
 const SearchResult = (props) => {
   
+  const [ flash, setFlash ] = useState("");
+  const [errorMessage, setErrorMessage] = useState(null);
+
+  const addToLibrary = (movie) => {
+    axios.post(('/movies'),{ 
+      title: movie.title,
+      overview: movie.overview,
+      release_date: movie.release_date,
+      image_url: movie.image_url,
+      external_id: movie.external_id
+    })
+      .then((response) => {
+        const flashMsg = movie.title + " is successfully added to the library! "
+        setFlash(flashMsg);
+        setTimeout(() => {
+          setFlash(null)
+        }, 2000);
+        props.updateMovie();
+      })
+      .catch((error) => {
+        setErrorMessage("Can't add the same movie twice!!");
+        setTimeout(() => {
+          setErrorMessage(null)
+        }, 2000);
+      });
+  }
+
   return (
     <section className="card movie-card">
       <div className="card--image">
@@ -17,7 +45,7 @@ const SearchResult = (props) => {
       </div>
       <button 
           className="btn btn-primary" 
-          onClick={() => {props.addToLibrary(props)} }
+          onClick={() => {addToLibrary(props)} }
           >
           Add To Library
       </button>
